Add a separate window widget for event list.
